+++ rustls/src/msgs/fragmenter.rs
@@ -1,5 +1,5 @@
+use crate::enums::ContentType;
 use crate::enums::ProtocolVersion;
-use crate::msgs::enums::ContentType;
 use crate::msgs::message::{BorrowedPlainMessage, PlainMessage};
 use crate::Error;
 pub const MAX_FRAGMENT_LEN: usize = 16384;
@@ -68,9 +68,9 @@ impl MessageFragmenter {
 #[cfg(test)]
 mod tests {
     use super::{MessageFragmenter, PACKET_OVERHEAD};
+    use crate::enums::ContentType;
     use crate::enums::ProtocolVersion;
     use crate::msgs::base::Payload;
-    use crate::msgs::enums::ContentType;
     use crate::msgs::message::{BorrowedPlainMessage, PlainMessage};
 
     fn msg_eq(
